+++ BetterChests/Framework/Models/Events/ItemTransferringEventArgs.cs
@@ -5,18 +5,13 @@ using StardewMods.Common.Services.Integrations.BetterChests.Interfaces;
 /// <inheritdoc cref="StardewMods.Common.Services.Integrations.BetterChests.Interfaces.IItemTransferring" />
 internal sealed class ItemTransferringEventArgs : EventArgs, IItemTransferring
 {
-    private bool isAllowed;
-    private bool isPrevented;
-
     /// <summary>Initializes a new instance of the <see cref="ItemTransferringEventArgs" /> class.</summary>
     /// <param name="into">The container being transferred into.</param>
     /// <param name="item">The item being transferred.</param>
-    /// <param name="isAllowedByDefault">Indicates whether the transfer is allowed by default.</param>
-    public ItemTransferringEventArgs(IStorageContainer into, Item item, bool isAllowedByDefault)
+    public ItemTransferringEventArgs(IStorageContainer into, Item item)
     {
         this.Into = into;
         this.Item = item;
-        this.IsAllowedByDefault = isAllowedByDefault;
     }
 
     /// <inheritdoc />
@@ -26,14 +21,14 @@ internal sealed class ItemTransferringEventArgs : EventArgs, IItemTransferring
     public Item Item { get; }
 
     /// <inheritdoc />
-    public bool IsAllowed => this.isAllowed && !this.isPrevented;
+    public bool IsAllowed { get; private set; }
 
     /// <inheritdoc />
-    public bool IsAllowedByDefault { get; }
+    public bool IsPrevented { get; private set; }
 
     /// <inheritdoc />
-    public void AllowTransfer() => this.isAllowed = true;
+    public void AllowTransfer() => this.IsAllowed = true;
 
     /// <inheritdoc />
-    public void PreventTransfer() => this.isPrevented = true;
+    public void PreventTransfer() => this.IsPrevented = true;
 }