+++ BetterChests/Framework/Services/Features/CategorizeChest.cs
@@ -5,7 +5,6 @@ using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
 using StardewMods.BetterChests.Framework.Interfaces;
 using StardewMods.BetterChests.Framework.Models.Events;
-using StardewMods.BetterChests.Framework.Services.Factory;
 using StardewMods.Common.Interfaces;
 using StardewMods.Common.Models;
 using StardewMods.Common.Services;
@@ -30,11 +29,11 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
 
     private readonly PerScreen<List<Item>> cachedItems = new(() => []);
     private readonly ICacheTable<ISearchExpression?> cachedSearches;
-    private readonly ContainerFactory containerFactory;
     private readonly PerScreen<ClickableTextureComponent> existingStacksButton;
     private readonly IInputHelper inputHelper;
     private readonly PerScreen<bool> isActive = new();
     private readonly ItemGrabMenuManager itemGrabMenuManager;
+    private readonly PerScreen<ClickableTextureComponent> rejectButton;
     private readonly PerScreen<ClickableTextureComponent> saveButton;
     private readonly SearchHandler searchHandler;
     private readonly PerScreen<string> searchText;
@@ -42,7 +41,6 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
     /// <summary>Initializes a new instance of the <see cref="CategorizeChest" /> class.</summary>
     /// <param name="assetHandler">Dependency used for handling assets.</param>
     /// <param name="cacheManager">Dependency used for managing cache tables.</param>
-    /// <param name="containerFactory">Dependency used for accessing containers.</param>
     /// <param name="eventManager">Dependency used for managing events.</param>
     /// <param name="inputHelper">Dependency used for checking and changing input state.</param>
     /// <param name="itemGrabMenuManager">Dependency used for managing the item grab menu.</param>
@@ -54,7 +52,6 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
     public CategorizeChest(
         AssetHandler assetHandler,
         CacheManager cacheManager,
-        ContainerFactory containerFactory,
         IEventManager eventManager,
         IInputHelper inputHelper,
         ItemGrabMenuManager itemGrabMenuManager,
@@ -66,11 +63,23 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
         : base(eventManager, log, manifest, modConfig)
     {
         this.cachedSearches = cacheManager.GetCacheTable<ISearchExpression?>();
-        this.containerFactory = containerFactory;
         this.inputHelper = inputHelper;
         this.itemGrabMenuManager = itemGrabMenuManager;
         this.searchHandler = searchHandler;
         this.searchText = searchText;
+
+        this.rejectButton = new PerScreen<ClickableTextureComponent>(
+            () => new ClickableTextureComponent(
+                new Rectangle(0, 0, 24, 24),
+                Game1.mouseCursors,
+                new Rectangle(322, 498, 12, 12),
+                2f)
+            {
+                name = "Reject",
+                hoverText = I18n.Button_RejectUncategorized_Name(),
+                myID = 8_675_308,
+            });
+
         this.saveButton = new PerScreen<ClickableTextureComponent>(
             () => new ClickableTextureComponent(
                 new Rectangle(0, 0, Game1.tileSize / 2, Game1.tileSize / 2),
@@ -78,7 +87,7 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
                 new Rectangle(142, 0, 16, 16),
                 2f)
             {
-                name = this.Id,
+                name = "Save",
                 hoverText = I18n.Button_SaveAsCategorization_Name(),
                 myID = 8_675_309,
             });
@@ -90,7 +99,7 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
                 new Rectangle(227, 425, 9, 9),
                 3f)
             {
-                name = this.Id,
+                name = "ExistingStacks",
                 hoverText = I18n.Button_IncludeExistingStacks_Name(),
                 myID = 8_675_310,
             });
@@ -139,6 +148,18 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
 
         var (mouseX, mouseY) = Game1.getMousePosition(true);
 
+        // Toggle reject uncategorized items
+        if (this.rejectButton.Value.containsPoint(mouseX, mouseY))
+        {
+            this.inputHelper.Suppress(e.Button);
+            container.Options.CategorizeChestBlockItems =
+                container.Options.CategorizeChestBlockItems == FeatureOption.Enabled
+                    ? FeatureOption.Disabled
+                    : FeatureOption.Enabled;
+
+            return;
+        }
+
         // Copy search text to categorization
         if (this.saveButton.Value.containsPoint(mouseX, mouseY))
         {
@@ -147,22 +168,20 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
             return;
         }
 
-        if (!this.existingStacksButton.Value.containsPoint(mouseX, mouseY))
-        {
-            return;
-        }
-
         // Toggle include existing stacks
-        this.inputHelper.Suppress(e.Button);
-        container.Options.CategorizeChestIncludeStacks =
-            container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled
-                ? FeatureOption.Disabled
-                : FeatureOption.Enabled;
+        if (this.existingStacksButton.Value.containsPoint(mouseX, mouseY))
+        {
+            this.inputHelper.Suppress(e.Button);
+            container.Options.CategorizeChestIncludeStacks =
+                container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled
+                    ? FeatureOption.Disabled
+                    : FeatureOption.Enabled;
 
-        this.existingStacksButton.Value.sourceRect =
-            container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled
-                ? new Rectangle(236, 425, 9, 9)
-                : new Rectangle(227, 425, 9, 9);
+            this.existingStacksButton.Value.sourceRect =
+                container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled
+                    ? new Rectangle(236, 425, 9, 9)
+                    : new Rectangle(227, 425, 9, 9);
+        }
     }
 
     private void OnItemGrabMenuChanged(ItemGrabMenuChangedEventArgs e)
@@ -195,6 +214,9 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
                 container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled
                     ? new Rectangle(236, 425, 9, 9)
                     : new Rectangle(227, 425, 9, 9);
+
+            this.rejectButton.Value.bounds.X = x + 64;
+            this.rejectButton.Value.bounds.Y = y + 2;
         }
         else
         {
@@ -204,17 +226,24 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
 
     private void OnRenderedActiveMenu(RenderedActiveMenuEventArgs e)
     {
-        if (!this.isActive.Value || this.itemGrabMenuManager.CurrentMenu is null)
+        var container = this.itemGrabMenuManager.Top.Container;
+        if (!this.isActive.Value || this.itemGrabMenuManager.CurrentMenu is null || container is null)
         {
             return;
         }
 
         var (mouseX, mouseY) = Game1.getMousePosition(true);
+
         this.saveButton.Value.tryHover(mouseX, mouseY);
         this.existingStacksButton.Value.tryHover(mouseX, mouseY);
+        this.rejectButton.Value.tryHover(mouseX, mouseY);
 
         this.saveButton.Value.draw(e.SpriteBatch);
         this.existingStacksButton.Value.draw(e.SpriteBatch);
+        this.rejectButton.Value.draw(
+            e.SpriteBatch,
+            container.Options.CategorizeChestBlockItems == FeatureOption.Enabled ? Color.White : Color.Gray,
+            1f);
 
         if (this.saveButton.Value.containsPoint(mouseX, mouseY))
         {
@@ -225,6 +254,12 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
         if (this.existingStacksButton.Value.containsPoint(mouseX, mouseY))
         {
             this.itemGrabMenuManager.CurrentMenu.hoverText = this.existingStacksButton.Value.hoverText;
+            return;
+        }
+
+        if (this.rejectButton.Value.containsPoint(mouseX, mouseY))
+        {
+            this.itemGrabMenuManager.CurrentMenu.hoverText = this.rejectButton.Value.hoverText;
         }
     }
 
@@ -300,39 +335,38 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
     private bool CanAcceptItem(IStorageContainer container, Item item, out bool accepted)
     {
         accepted = false;
-        if (container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled)
+        var includeStacks = container.Options.CategorizeChestIncludeStacks == FeatureOption.Enabled;
+        var hasStacks = container.Items.ContainsId(item.QualifiedItemId);
+        if (includeStacks && hasStacks)
         {
-            accepted = container.Items.ContainsId(item.QualifiedItemId);
+            accepted = true;
             return true;
         }
 
+        // Cannot handle if there is no search term
         if (string.IsNullOrWhiteSpace(container.Options.CategorizeChestSearchTerm))
         {
             return false;
         }
 
-        // Return results from cache
-        if (this.cachedSearches.TryGetValue(container.Options.CategorizeChestSearchTerm, out var searchExpression))
+        // Retrieve search expression from cache or generate a new one
+        if (!this.cachedSearches.TryGetValue(container.Options.CategorizeChestSearchTerm, out var searchExpression))
         {
-            if (searchExpression is null)
-            {
-                return false;
-            }
-
-            accepted = searchExpression.PartialMatch(item);
-            return true;
+            this.cachedSearches.AddOrUpdate(
+                container.Options.CategorizeChestSearchTerm,
+                this.searchHandler.TryParseExpression(container.Options.CategorizeChestSearchTerm, out searchExpression)
+                    ? searchExpression
+                    : null);
         }
 
-        // Parse search term and cache results
-        if (this.searchHandler.TryParseExpression(container.Options.CategorizeChestSearchTerm, out searchExpression))
+        // Cannot handle if search term is invalid
+        if (searchExpression is null)
         {
-            this.cachedSearches.AddOrUpdate(container.Options.CategorizeChestSearchTerm, searchExpression);
-            accepted = searchExpression.PartialMatch(item);
-            return true;
+            return false;
         }
 
-        // Cache null result if search term fails to parse
-        this.cachedSearches.AddOrUpdate(container.Options.CategorizeChestSearchTerm, null);
-        return false;
+        // Check if item matches search expressions
+        accepted = searchExpression.PartialMatch(item);
+        return true;
     }
 }