+++ BetterChests/Framework/Services/Features/StashToChest.cs
@@ -11,7 +11,6 @@ using StardewMods.Common.Services.Integrations.BetterChests.Interfaces;
 using StardewMods.Common.Services.Integrations.FauxCore;
 using StardewMods.Common.Services.Integrations.ToolbarIcons;
 using StardewValley.Locations;
-using StardewValley.Menus;
 
 /// <summary>Stash items into placed chests and chests in the farmer's inventory.</summary>
 internal sealed class StashToChest : BaseFeature<StashToChest>
@@ -20,6 +19,7 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
     private readonly ContainerFactory containerFactory;
     private readonly ContainerHandler containerHandler;
     private readonly IInputHelper inputHelper;
+    private readonly ItemGrabMenuManager itemGrabMenuManager;
     private readonly ToolbarIconsIntegration toolbarIconsIntegration;
 
     /// <summary>Initializes a new instance of the <see cref="StashToChest" /> class.</summary>
@@ -28,6 +28,7 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
     /// <param name="containerHandler">Dependency used for handling operations between containers.</param>
     /// <param name="eventManager">Dependency used for managing events.</param>
     /// <param name="inputHelper">Dependency used for checking and changing input state.</param>
+    /// <param name="itemGrabMenuManager">Dependency used for managing the item grab menu.</param>
     /// <param name="log">Dependency used for logging debug information to the console.</param>
     /// <param name="manifest">Dependency for accessing mod manifest.</param>
     /// <param name="modConfig">Dependency used for accessing config data.</param>
@@ -38,6 +39,7 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
         ContainerHandler containerHandler,
         IEventManager eventManager,
         IInputHelper inputHelper,
+        ItemGrabMenuManager itemGrabMenuManager,
         ILog log,
         IManifest manifest,
         IModConfig modConfig,
@@ -48,6 +50,7 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
         this.containerFactory = containerFactory;
         this.containerHandler = containerHandler;
         this.inputHelper = inputHelper;
+        this.itemGrabMenuManager = itemGrabMenuManager;
         this.toolbarIconsIntegration = toolbarIconsIntegration;
     }
 
@@ -60,6 +63,7 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
         // Events
         this.Events.Subscribe<ButtonsChangedEventArgs>(this.OnButtonsChanged);
         this.Events.Subscribe<ButtonPressedEventArgs>(this.OnButtonPressed);
+        this.Events.Subscribe<RenderingActiveMenuEventArgs>(this.OnRenderingActiveMenu);
 
         // Integrations
         if (!this.toolbarIconsIntegration.IsLoaded)
@@ -82,6 +86,7 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
         // Events
         this.Events.Unsubscribe<ButtonsChangedEventArgs>(this.OnButtonsChanged);
         this.Events.Unsubscribe<ButtonPressedEventArgs>(this.OnButtonPressed);
+        this.Events.Unsubscribe<RenderingActiveMenuEventArgs>(this.OnRenderingActiveMenu);
 
         // Integrations
         if (!this.toolbarIconsIntegration.IsLoaded)
@@ -96,21 +101,61 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
     private void OnButtonPressed(ButtonPressedEventArgs e)
     {
         if (e.Button is not SButton.MouseLeft
+            || this.itemGrabMenuManager.CurrentMenu?.fillStacksButton is null
+            || this.itemGrabMenuManager.Bottom.Container is null
+            || this.itemGrabMenuManager.Top.Container is null
             || !this.containerFactory.TryGetOne(out var container)
             || container.Options.StashToChest is RangeOption.Disabled or RangeOption.Default)
         {
             return;
         }
 
-        var (x, y) = Game1.getMousePosition(true);
-        if ((Game1.activeClickableMenu as ItemGrabMenu)?.fillStacksButton?.containsPoint(x, y) != true)
+        var (mouseX, mouseY) = Game1.getMousePosition(true);
+        if (!this.itemGrabMenuManager.CurrentMenu.fillStacksButton.containsPoint(mouseX, mouseY))
         {
             return;
         }
 
         this.inputHelper.Suppress(e.Button);
-        this.StashIntoContainer(container);
         Game1.playSound("Ship");
+
+        // Stash to existing stacks only
+        if (!this.Config.Controls.TransferItems.IsDown()
+            || this.itemGrabMenuManager.Top.Container.Options.StashToChest is RangeOption.Disabled
+                or RangeOption.Default)
+        {
+            this.containerHandler.Transfer(
+                this.itemGrabMenuManager.Bottom.Container,
+                this.itemGrabMenuManager.Top.Container,
+                out _,
+                existingOnly: true);
+
+            return;
+        }
+
+        // Stash using categorization rules
+        var (from, to) = this.Config.Controls.TransferItemsReverse.IsDown()
+            ? (this.itemGrabMenuManager.Top.Container, this.itemGrabMenuManager.Bottom.Container)
+            : (this.itemGrabMenuManager.Bottom.Container, this.itemGrabMenuManager.Top.Container);
+
+        if (!this.containerHandler.Transfer(from, to, out var amounts))
+        {
+            return;
+        }
+
+        foreach (var (name, amount) in amounts)
+        {
+            if (amount > 0)
+            {
+                this.Log.Trace(
+                    "{0}: {{ Item: {1}, Quantity: {2}, From: {3}, To: {4} }}",
+                    this.Id,
+                    name,
+                    amount,
+                    from,
+                    to);
+            }
+        }
     }
 
     private void OnButtonsChanged(ButtonsChangedEventArgs e)
@@ -138,6 +183,40 @@ internal sealed class StashToChest : BaseFeature<StashToChest>
         this.inputHelper.SuppressActiveKeybinds(this.Config.Controls.StashItems);
         this.StashIntoContainer(container);
         Game1.playSound("Ship");
+    }
+
+    private void OnRenderingActiveMenu(RenderingActiveMenuEventArgs obj)
+    {
+        if (this.itemGrabMenuManager.CurrentMenu?.fillStacksButton is null
+            || !this.containerFactory.TryGetOne(out var container)
+            || container.Options.StashToChest is RangeOption.Disabled or RangeOption.Default)
+        {
+            return;
+        }
+
+        var (mouseX, mouseY) = Game1.getMousePosition(true);
+        if (!this.Config.Controls.TransferItems.IsDown()
+            || !this.itemGrabMenuManager.CurrentMenu.fillStacksButton.containsPoint(mouseX, mouseY))
+        {
+            this.itemGrabMenuManager.CurrentMenu.fillStacksButton.texture = Game1.mouseCursors;
+            this.itemGrabMenuManager.CurrentMenu.fillStacksButton.sourceRect = new Rectangle(103, 469, 16, 16);
+            this.itemGrabMenuManager.CurrentMenu.fillStacksButton.hoverText =
+                Game1.content.LoadString("Strings\\UI:ItemGrab_FillStacks");
+
+            return;
+        }
+
+        this.itemGrabMenuManager.CurrentMenu.fillStacksButton.texture = this.assetHandler.Icons.Value;
+
+        this.itemGrabMenuManager.CurrentMenu.fillStacksButton.sourceRect =
+            this.Config.Controls.TransferItemsReverse.IsDown()
+                ? new Rectangle(96, 0, 16, 16)
+                : new Rectangle(80, 0, 16, 16);
+
+        this.itemGrabMenuManager.CurrentMenu.fillStacksButton.hoverText =
+            this.Config.Controls.TransferItemsReverse.IsDown()
+                ? I18n.Button_TransferDown_Name()
+                : I18n.Button_TransferUp_Name();
     }
 
     private void OnIconPressed(IIconPressedEventArgs e)